test(matrix): restructure tests for matrix IO, media, invites, and relay
- remove obsolete matrix relay tests from tests/test_matrix_utils.py
- add image upload and sending image tests in tests/test_matrix_utils_media.py
- use MATRIX_EVENT_TYPE_ROOM_MESSAGE constant in tests
- adjust relay tests to reflect updated matrix_utils API and mocks
- drop path-hack in invites tests for cleaner imports
